+++ CHANGELOG.md
@@ -139,6 +139,7 @@ You can find its changes [documented below](#070---2021-01-01).
 - X11: window focus events ([#1938] by [@Maan2003]
 - Preserve the aspect ratio of a clipped region in an Image ([#2195] by [@barsae])
 - GTK: Hot state now properly resets when the mouse leaves the window via an occluded part. ([#2324] by [@xStrom])
+- Windows: generate `commands::OPEN_FILES` when using multiselect in file open dialog. ([#2332] by [@ppykalo])
 
 ### Visual
 
@@ -578,6 +579,7 @@ Last release without a changelog :(
 [@benoitryder]: https://github.com/benoitryder
 [@sprocklem]: https://github.com/sprocklem
 [@cbondurant]: https://github.com/cbondurant
+[@ppykalo]: https://github.com/ppykalo
 
 [#599]: https://github.com/linebender/druid/pull/599
 [#611]: https://github.com/linebender/druid/pull/611
@@ -884,6 +886,7 @@ Last release without a changelog :(
 [#2323]: https://github.com/linebender/druid/pull/2323
 [#2324]: https://github.com/linebender/druid/pull/2324
 [#2331]: https://github.com/linebender/druid/pull/2331
+[#2332]: https://github.com/linebender/druid/pull/2332
 
 [Unreleased]: https://github.com/linebender/druid/compare/v0.7.0...master
 [0.7.0]: https://github.com/linebender/druid/compare/v0.6.0...v0.7.0